Fix typescript
