+++ client/src/components/cart/Cart.tsx
@@ -1,7 +1,11 @@
 import React from "react";
 
 const Cart = (): JSX.Element => {
-  return <div>Cart</div>;
+  return (
+    <div>
+      Cart
+    </div>
+  )
 };
 
 export default Cart;
