+++ client/src/components/products/cards/Cards.tsx
@@ -9,7 +9,7 @@ import ReactPaginate from "react-paginate";
 import { useDispatch, useSelector } from "react-redux";
 import { State } from "../../../redux/reducers/index";
 import { getProducts } from "../../../redux/actions/products";
-import { Product } from "../../../redux/interface";
+import { Data_Paginate, FILTER_BOX, Product, Subcategory } from "../../../redux/interface";
 import Loading from "../../loading/Loading";
 import Categories from "../categories/Categories";
 import { ProductsContainer } from "../ProductsStyles";
@@ -17,19 +17,10 @@ import NotFound from "../../notFound/NotFound";
 import { chargeFilter, filterProducts, removeFilter, resetFilterProducts } from "../../../redux/actions/filterByCategory";
 import { execPath } from "process";
 import { filterByBrand } from "../../../redux/actions/filterByBrand";
-export interface IData {
-  length: number;
-  page: (numberOfPage: number) => void;
-}
 
 export interface ORDER {
   page: (numberOfPage: number) => void;
   orders: (typeorder: string) => void;
-}
-
-export interface FILTER_BOX {
-  subcategory: string,
-  brand: string
 }
 
 const Cards = (): JSX.Element => {
@@ -45,20 +36,17 @@ const Cards = (): JSX.Element => {
   const copyProductsList = useSelector((state: State) => state.products.copyProducts)
   const filteredProductList = useSelector((state: State) => state.filteredProducts.filteredProducts);
   const allSubcategories = useSelector((state: State) => state.categories.subcategories);
-  const notFound = useSelector((state: State) => state.products.not_found);
-  const pageState = useSelector((state: State) => state.page);
 
   const page = (numberOfPage: number): void => {
     setCurrentPage(numberOfPage);
   };
   const orders = (typeorder: string): void => {
     if (typeorder !== 'asc-price order' && typeorder !== 'des-price order' && typeorder !== 'des-name order' && typeorder !== 'asc-name order' && typeorder !== 'Order by order') {
-      let existCat = allSubcategories.filter((e: any) => e.name === typeorder)
+      let existCat = allSubcategories.filter((e: Subcategory) => e.name === typeorder)
       if (existCat.length === 1) {
         setFilterBox({ ...filterBox, subcategory: typeorder })
         dispatch(chargeFilter(copyProductsList))
         if (filterBox.brand.length !== 0) {
-          console.log('Hay brand')
           dispatch(filterByBrand(filterBox.brand))
         }
         dispatch(filterProducts(typeorder))
@@ -67,7 +55,6 @@ const Cards = (): JSX.Element => {
         setFilterBox({ ...filterBox, brand: typeorder })
         dispatch(chargeFilter(copyProductsList))
         if (filterBox.subcategory.length !== 0) {
-          console.log('Hay subcategory')
           dispatch(filterProducts(filterBox.subcategory))
         }
         dispatch(filterByBrand(typeorder))
@@ -90,26 +77,22 @@ const Cards = (): JSX.Element => {
   const finalProduct = currentPage * 32;
   const firstProduct = finalProduct - 32;
   let newProductsList: Product[] = [];
-  let auxiliar: string = "Load";
-  // (newProductsList = productsList.slice(firstProduct, finalProduct));
-  // let newProductsList: Product[] = [];
   filteredProductList.length > 0
     ? (newProductsList = filteredProductList.slice(firstProduct, finalProduct))
     : (newProductsList = productsList.slice(firstProduct, finalProduct))
 
   // implementing react paginate
 
-  const handlePageClick = (data: any) => {
+  const handlePageClick = (data: Data_Paginate): void => {
     setCurrentPage(data.selected + 1);
   };
 
-  const resetFilter = (e: any): void => {
-    e.preventDefault()
+  const resetFilter = (e: string): void => {
     if (filterBox.subcategory.length === 0 || filterBox.brand.length === 0) {
       dispatch(chargeFilter(copyProductsList))
-    } else if (filterBox.subcategory === e.target.value) dispatch(removeFilter(filterBox.brand))
+    } else if (filterBox.subcategory === e) dispatch(removeFilter(filterBox.brand))
     else dispatch(removeFilter(filterBox.subcategory))
-    let existCat = allSubcategories.filter((s: any) => s.name === e.target.value)
+    let existCat = allSubcategories.filter((s: any) => s.name === e)
     if (existCat.length === 0) setFilterBox({ ...filterBox, brand: "" })
     else setFilterBox({ ...filterBox, subcategory: "" })
   }
@@ -137,8 +120,8 @@ const Cards = (): JSX.Element => {
               :
               filteredProductList.length > 0 ?
                 <>
-                  {filterBox.subcategory.length !== 0 ? <span><button value={filterBox.subcategory} onClick={(e) => resetFilter(e)} className="btn btn-primary mt-2 mr-2">{filterBox.subcategory}</button></span> : ""}
-                  {filterBox.brand.length !== 0 ? <span><button value={filterBox.brand} onClick={(e) => resetFilter(e)} className="btn btn-primary mt-2 mr-2">{filterBox.brand}</button></span> : ""}
+                  {filterBox.subcategory.length !== 0 ? <span><button onClick={() => resetFilter(filterBox.subcategory)} className="btn btn-primary mt-2 mr-2">{filterBox.subcategory}</button></span> : ""}
+                  {filterBox.brand.length !== 0 ? <span><button onClick={() => resetFilter(filterBox.brand)} className="btn btn-primary mt-2 mr-2">{filterBox.brand}</button></span> : ""}
                   <div className="mt-3 row row-cols-1 row-cols-sm-2 row-cols-md-3 row-cols-xxl-4 g-4 d-flex justify-content-center">
                     {newProductsList.map((e: Product) => {
                       return (
