+++ client/src/components/products/cards/card/Card.tsx
@@ -1,7 +1,8 @@
-import React from "react";
+import React, { useState } from "react";
 import { CardComponent, CardFooter, ProductIMG } from "./CardStyles";
 import cartIcon from "../../../../icons/cart-icon.png";
 import { Link } from "react-router-dom";
+import { useDispatch } from "react-redux";
 
 interface props {
   name: string;
@@ -20,13 +21,13 @@ const Card = ({ name, image, price, id }: props) => {
             alt={image}
             className="card-img-top"
           ></ProductIMG>
-          {name.length > 40 ? <p className="card-title m-2">{name.slice(0,40)}...</p> : <p className="card-title m-2">{name}</p>}
+          {name.length > 40 ? <p className="card-title m-2">{name.slice(0, 40)}...</p> : <p className="card-title m-2">{name}</p>}
         </Link>
       </div>
       <CardFooter className="card-footer d-flex align-items-end justify-content-between">
         <h5 className="m-3">${price}</h5>
         <button type="button" className="btn btn-primary h-100">
-          <img src={cartIcon} alt=""></img>
+          <img src={cartIcon} alt="" ></img>
         </button>
       </CardFooter>
     </CardComponent>
